анализ/requirements.txt
python-telegram-bot==20.5
sqlalchemy==2.0.23
python-dateutil==2.8.2